refactor: extract basic types
