+++ src/constants.ts
@@ -1,6 +1,6 @@
 import {
-  CardanoEra, Network, NetworkIds, ProtocolMagics,
-} from './types'
+  Network, NetworkIds, ProtocolMagics, CardanoEra,
+} from './basicTypes'
 import { invertObject } from './util'
 
 // taken from https://book.world.dev.cardano.org/environments.html
