+++ test/integration/trezor/node/signingFiles.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-len */
 
 import { HARDENED_THRESHOLD } from '../../../../src/constants'
-import { HwSigningType } from '../../../../src/types'
+import { HwSigningType } from '../../../../src/basicTypes'
 
 // mnemonic "all all all all all all all all all all all all"
